requirements.txt
fastapi==0.63.0
uvicorn[standard]==0.13.4
gino==1.0.1
gino-starlette==0.1.1
requests==2.25.1
click==7.1.2
aiounittest==1.4.0
locust==1.4.4